Fixes an issue where end could be < start
This also removes some special casing when rest is an empty string,
simplifying some of the code.

The issue was best highlighted when we had empty strings - the end
would point to the start and the start would point to the end, it worked
out basically accidentally. But now it's fixed anyway.