Introduces object depth to parse array element
I think when we are parsing arrays we will need to track the depth of an
object so we can emit errors etc as expected. But to do that we need to
track both an array depth and an object depth. In a more sophisticated
data type we'd have to have one depth tracker per collection type that
can be nested inside another. Luckily JSON only has two collection types
objects and arrays so this is manageable, if a bit ugly.